pyproject.toml
[project]
dependencies = [
    "browser-cookie3>=0.20.1",
    "click>=8.3.1",
    "httpx>=0.28.1",
    "junitparser>=4.0.2",
    "py-markdown-table>=1.3.0",
    "pyrage>=1.3.0",
    "pyobjc-framework-LaunchServices>=11.0; sys_platform == 'darwin'",
    "tomli>=2.3.0",
    "tomli-w>=1.2.0",
]
description = "Utility CLI for interacting with Advent of Code"
name = "x"
requires-python = ">=3.14"
version = "0.1.0"

[dependency-groups]
dev = ["ruff>=0.14.8", "pyrage-stubs>=1.3.1"]
